+++ Starward.Core/Hyperion/StarRail/Ledger/LedgerDetail.cs
@@ -10,7 +10,7 @@ public class LedgerDetail : IJsonOnDeserialized
 
     [JsonPropertyName("uid")]
     [JsonNumberHandling(JsonNumberHandling.AllowReadingFromString | JsonNumberHandling.WriteAsString)]
-    public int Uid { get; set; }
+    public long Uid { get; set; }
 
     [JsonPropertyName("region")]
     public string Region { get; set; }
